+++ docs/guides/external-node/00_quick_start.md
@@ -34,8 +34,7 @@ cd docker-compose-examples
 docker compose --file testnet-external-node-docker-compose.yml down --volumes
 ```
 
-You can see the status of the node (after recovery) in
-[local grafana dashboard](http://localhost:3000/dashboards).
+You can see the status of the node (after recovery) in [local grafana dashboard](http://localhost:3000/dashboards).
 
 Those commands start ZKsync node locally inside docker.
