+++ zk_toolbox/crates/zk_inception/src/commands/chain/args/mod.rs
@@ -2,3 +2,4 @@ pub mod build_transactions;
 pub mod create;
 pub mod genesis;
 pub mod init;
+pub mod init_configs;
