+++ zk_toolbox/crates/zk_inception/src/commands/chain/init.rs
@@ -1,36 +1,30 @@
-use anyhow::{bail, Context};
+use anyhow::Context;
 use common::{config::global_config, git, logger, spinner::Spinner};
-use config::{
-    copy_configs, ports_config, set_l1_rpc_url, traits::SaveConfigWithBasePath,
-    update_from_chain_config, update_ports, ChainConfig, ContractsConfig, EcosystemConfig,
-    GeneralConfig,
-};
-use ethers::types::Address;
+use config::{traits::SaveConfigWithBasePath, ChainConfig, EcosystemConfig};
 use types::BaseToken;
 use xshell::Shell;
 
 use super::common::{distribute_eth, mint_base_token};
 use crate::{
     accept_ownership::accept_admin,
-    commands::{
-        chain::{
-            args::init::{InitArgs, InitArgsFinal},
-            deploy_l2_contracts, deploy_paymaster,
-            genesis::genesis,
-            register_chain::register_chain,
-            set_token_multiplier_setter::set_token_multiplier_setter,
-            setup_legacy_bridge::setup_legacy_bridge,
+    commands::chain::{
+        args::{
+            init::{InitArgs, InitArgsFinal},
+            init_configs::InitConfigsArgsFinal,
         },
-        portal::update_portal_config,
+        deploy_l2_contracts, deploy_paymaster,
+        genesis_all::genesis,
+        init_configs::init_configs,
+        register_chain::register_chain,
+        set_token_multiplier_setter::set_token_multiplier_setter,
+        setup_legacy_bridge::setup_legacy_bridge,
     },
     messages::{
-        msg_initializing_chain, MSG_ACCEPTING_ADMIN_SPINNER, MSG_CHAIN_CONFIGS_INITIALIZED,
-        MSG_CHAIN_INITIALIZED, MSG_CHAIN_NOT_FOUND_ERR, MSG_DEPLOYING_PAYMASTER,
-        MSG_GENESIS_DATABASE_ERR, MSG_PORTAL_FAILED_TO_CREATE_CONFIG_ERR, MSG_PORTS_CONFIG_ERR,
+        msg_initializing_chain, MSG_ACCEPTING_ADMIN_SPINNER, MSG_CHAIN_INITIALIZED,
+        MSG_CHAIN_NOT_FOUND_ERR, MSG_DEPLOYING_PAYMASTER, MSG_GENESIS_DATABASE_ERR,
         MSG_REGISTERING_CHAIN_SPINNER, MSG_SELECTED_CONFIG,
         MSG_UPDATING_TOKEN_MULTIPLIER_SETTER_SPINNER, MSG_WALLET_TOKEN_MULTIPLIER_SETTER_NOT_FOUND,
     },
-    utils::consensus::{generate_consensus_keys, get_consensus_config, get_consensus_secrets},
 };
 
 pub(crate) async fn run(args: InitArgs, shell: &Shell) -> anyhow::Result<()> {
@@ -47,25 +41,22 @@ pub(crate) async fn run(args: InitArgs, shell: &Shell) -> anyhow::Result<()> {
 
     init(&mut args, shell, &config, &chain_config).await?;
 
-    if args.configs_only {
-        logger::success(MSG_CHAIN_CONFIGS_INITIALIZED);
-    } else {
-        logger::success(MSG_CHAIN_INITIALIZED);
-    }
+    logger::success(MSG_CHAIN_INITIALIZED);
     Ok(())
 }
 
 pub async fn init(
-    init_args: &mut InitArgsFinal,
+    init_args: &InitArgsFinal,
     shell: &Shell,
     ecosystem_config: &EcosystemConfig,
     chain_config: &ChainConfig,
 ) -> anyhow::Result<()> {
-    let mut contracts_config = init_configs(init_args, shell, ecosystem_config, chain_config)?;
-    if init_args.configs_only {
-        return Ok(());
-    }
+    // Initialize configs
+    let init_configs_args = InitConfigsArgsFinal::from_chain_init_args(&init_args);
+    let mut contracts_config =
+        init_configs(&init_configs_args, shell, ecosystem_config, chain_config).await?;
 
+    // Fund some wallet addresses with ETH or base token (only for Localhost)
     distribute_eth(ecosystem_config, chain_config, init_args.l1_rpc_url.clone()).await?;
     mint_base_token(ecosystem_config, chain_config, init_args.l1_rpc_url.clone()).await?;
 
@@ -162,60 +153,6 @@ pub async fn init(
     genesis(init_args.genesis_args.clone(), shell, chain_config)
         .await
         .context(MSG_GENESIS_DATABASE_ERR)?;
-
-    update_portal_config(shell, chain_config)
-        .await
-        .context(MSG_PORTAL_FAILED_TO_CREATE_CONFIG_ERR)?;
-
-    Ok(())
-}
-
-fn init_configs(
-    init_args: &mut InitArgsFinal,
-    shell: &Shell,
-    ecosystem_config: &EcosystemConfig,
-    chain_config: &ChainConfig,
-) -> anyhow::Result<ContractsConfig> {
-    copy_configs(shell, &ecosystem_config.link_to_code, &chain_config.configs)?;
-
-    let mut general_config = chain_config.get_general_config()?;
-    apply_port_offset(init_args.port_offset, &mut general_config)?;
-    let ports = ports_config(&general_config).context(MSG_PORTS_CONFIG_ERR)?;
-
-    let consensus_keys = generate_consensus_keys();
-    let consensus_config =
-        get_consensus_config(chain_config, ports, Some(consensus_keys.clone()), None)?;
-    general_config.consensus_config = Some(consensus_config);
-    general_config.save_with_base_path(shell, &chain_config.configs)?;
-
-    let mut genesis_config = chain_config.get_genesis_config()?;
-    update_from_chain_config(&mut genesis_config, chain_config);
-    genesis_config.save_with_base_path(shell, &chain_config.configs)?;
-
-    // Copy ecosystem contracts
-    let mut contracts_config = ecosystem_config.get_contracts_config()?;
-    contracts_config.l1.diamond_proxy_addr = Address::zero();
-    contracts_config.l1.governance_addr = Address::zero();
-    contracts_config.l1.chain_admin_addr = Address::zero();
-    contracts_config.l1.base_token_addr = chain_config.base_token.address;
-    contracts_config.save_with_base_path(shell, &chain_config.configs)?;
-
-    let mut secrets = chain_config.get_secrets_config()?;
-    set_l1_rpc_url(&mut secrets, init_args.l1_rpc_url.clone())?;
-    secrets.consensus = Some(get_consensus_secrets(&consensus_keys));
-    secrets.save_with_base_path(shell, &chain_config.configs)?;
-
-    Ok(contracts_config)
-}
-
-fn apply_port_offset(port_offset: u16, general_config: &mut GeneralConfig) -> anyhow::Result<()> {
-    let Some(mut ports_config) = ports_config(general_config) else {
-        bail!("Missing ports config");
-    };
-
-    ports_config.apply_offset(port_offset);
-
-    update_ports(general_config, &ports_config)?;
 
     Ok(())
 }
